pyproject.toml
[project]
name = "live-trading-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alpaca_trade_api",
    "numpy",
    "requests",
    "python-dotenv",
    "toml",
]